update, refactor not functional
