+++ Backend/PreMatricula/src/main/java/com/example/preMatricula/repositories/DisciplinaRepository.java
@@ -1,4 +1,4 @@
-package com.example.preMatricula.interfaces;
+package com.example.preMatricula.repositories;
 
 import com.example.preMatricula.entities.Discipline;
