+++ Backend/PreMatricula/src/main/java/com/example/preMatricula/services/UserService.java
@@ -8,7 +8,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import com.example.preMatricula.entities.User;
-import com.example.preMatricula.interfaces.UserRepository;
+import com.example.preMatricula.repositories.UserRepository;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseToken;
